fix(toxav): harden video processing and fix large frame handling
- Remove invalid vpx_img_free call in vc_iterate.

- Make vc_reconfigure_encoder atomic and add resolution limits.

- Support 32-bit RTPMessage lengths to correctly handle large frames.

- Ensure len is correctly set for assembled video frames.

- Add security checks for incoming frame sizes and resolutions.

- Significantly expand video and rtp unit tests covering security scenarios.

- Clean up VCSession structure and improve internal API consistency.

- Use #define for security constants in video module.

+++ toxav/audio.c
@@ -194,7 +194,7 @@ void ac_iterate(ACSession *ac)
             }
         } else {
             const uint8_t *msg_data = rtp_message_data(msg);
-            const uint16_t msg_length = rtp_message_len(msg);
+            const uint32_t msg_length = rtp_message_len(msg);
 
             if (msg_length <= 4) {
                 LOGGER_WARNING(ac->log, "Packet too short: %u", msg_length);
